refactor(skeleton): split OriginalSkeletonTree impls
