+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree.rs
@@ -1,3 +1,4 @@
+pub mod compute_updated_skeleton_tree;
 pub mod create_tree;
 pub mod errors;
 pub mod node;
