+++ crates/committer/src/patricia_merkle_tree/original_skeleton_tree/create_tree.rs
@@ -6,10 +6,9 @@ use crate::patricia_merkle_tree::node_data::inner_node::EdgeData;
 use crate::patricia_merkle_tree::node_data::inner_node::NodeData;
 use crate::patricia_merkle_tree::node_data::inner_node::PathToBottom;
 use crate::patricia_merkle_tree::node_data::leaf::LeafDataImpl;
-use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTree;
+use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTreeImpl;
 use crate::patricia_merkle_tree::original_skeleton_tree::tree::OriginalSkeletonTreeResult;
 use crate::patricia_merkle_tree::types::TreeHeight;
-use crate::patricia_merkle_tree::updated_skeleton_tree::tree::UpdatedSkeletonTreeImpl;
 use crate::patricia_merkle_tree::{
     original_skeleton_tree::node::OriginalSkeletonNode, types::NodeIndex,
 };
@@ -25,12 +24,6 @@ use std::collections::HashMap;
 #[cfg(test)]
 #[path = "create_tree_test.rs"]
 pub mod create_tree_test;
-
-#[allow(dead_code)]
-pub(crate) struct OriginalSkeletonTreeImpl {
-    pub(crate) nodes: HashMap<NodeIndex, OriginalSkeletonNode<LeafDataImpl>>,
-    tree_height: TreeHeight,
-}
 
 struct SubTree<'a> {
     pub sorted_leaf_indices: &'a [NodeIndex],
@@ -215,45 +208,7 @@ impl OriginalSkeletonTreeImpl {
         Ok(subtrees_roots)
     }
 
-    fn get_node_height(&self, index: &NodeIndex) -> TreeHeight {
-        TreeHeight(self.tree_height.0 - index.bit_length() + 1)
-    }
-
-    /// Returns whether a root of a subtree has leaves on both sides. Assumes:
-    /// * The leaf indices array is sorted.
-    /// * All leaves are descendants of the root.
-    // TODO(Tzahi, 10/5/2024): Move implementation to a separate module.
-    #[allow(dead_code)]
-    fn has_leaves_on_both_sides(&self, root_index: &NodeIndex, leaf_indices: &[NodeIndex]) -> bool {
-        if leaf_indices.is_empty() {
-            return false;
-        }
-
-        let root_height = self.get_node_height(root_index);
-        let assert_child = |leaf_index: NodeIndex| {
-            if (leaf_index >> root_height.0) != *root_index {
-                panic!("Leaf is not a descendant of the root.");
-            }
-        };
-
-        let first_leaf = leaf_indices[0];
-        assert_child(first_leaf);
-        if leaf_indices.len() == 1 {
-            return false;
-        }
-
-        let last_leaf = leaf_indices
-            .last()
-            .expect("leaf_indices unexpectedly empty.");
-        assert_child(*last_leaf);
-
-        let child_direction_mask = U256::ONE << (root_height.0 - 1);
-        (first_leaf.0 & child_direction_mask) != (last_leaf.0 & child_direction_mask)
-    }
-}
-
-impl OriginalSkeletonTree<LeafDataImpl> for OriginalSkeletonTreeImpl {
-    fn create_tree(
+    pub(crate) fn create_tree_impl(
         storage: &impl Storage,
         sorted_leaf_indices: &[NodeIndex],
         root_hash: HashOutput,
@@ -270,13 +225,5 @@ impl OriginalSkeletonTree<LeafDataImpl> for OriginalSkeletonTreeImpl {
         };
         skeleton_tree.fetch_nodes(vec![main_subtree], storage)?;
         Ok(skeleton_tree)
-    }
-
-    #[allow(dead_code)]
-    fn compute_updated_skeleton_tree(
-        &self,
-        _index_to_updated_leaf: HashMap<NodeIndex, LeafDataImpl>,
-    ) -> OriginalSkeletonTreeResult<UpdatedSkeletonTreeImpl<LeafDataImpl>> {
-        todo!()
     }
 }
